refactor: Restructure project for better organization and maintainability
