+++ __init__.py
@@ -1,17 +1,17 @@
 """
 ComfyUI_Wan - A custom node for ComfyUI that integrates Wan models
-for text-to-image, image-to-video, and text-to-video generation.
+for text-to-image, image-to-video, text-to-video, first-last-frame-tovideo, and VACE generation.
 """
 
-from .wan_t2i import WanT2IGenerator
-from .wan_i2v import WanI2VGenerator
-from .wan_t2v import WanT2VGenerator
-from .wan_ii2v import WanII2VGenerator
-from .wan_vace_image_reference import WanVACEImageReference
-from .wan_vace_video_repainting import WanVACEVideoRepainting
-from .wan_vace_video_edit import WanVACEVideoEdit
-from .wan_vace_video_extension import WanVACEVideoExtension
-from .wan_vace_video_outpainting import WanVACEVideoOutpainting
+from .generators.t2i import WanT2IGenerator
+from .generators.i2v import WanI2VGenerator
+from .generators.t2v import WanT2VGenerator
+from .generators.ii2v import WanII2VGenerator
+from .vace.image_reference import WanVACEImageReference
+from .vace.video_repainting import WanVACEVideoRepainting
+from .vace.video_edit import WanVACEVideoEdit
+from .vace.video_extension import WanVACEVideoExtension
+from .vace.video_outpainting import WanVACEVideoOutpainting
 
 NODE_CLASS_MAPPINGS = {
     "WanT2IGenerator": WanT2IGenerator,
