+++ core/__init__.py
@@ -1,0 +1,3 @@
+from .base import WanAPIBase, COMFYUI_AVAILABLE
+
+__all__ = ['WanAPIBase', 'COMFYUI_AVAILABLE']
