+++ generators/__init__.py
@@ -1,0 +1,6 @@
+from .t2i import WanT2IGenerator
+from .i2v import WanI2VGenerator
+from .t2v import WanT2VGenerator
+from .ii2v import WanII2VGenerator
+
+__all__ = ['WanT2IGenerator', 'WanI2VGenerator', 'WanT2VGenerator', 'WanII2VGenerator']
