+++ vace/__init__.py
@@ -1,0 +1,8 @@
+from .image_reference import WanVACEImageReference
+from .video_repainting import WanVACEVideoRepainting
+from .video_edit import WanVACEVideoEdit
+from .video_extension import WanVACEVideoExtension
+from .video_outpainting import WanVACEVideoOutpainting
+
+__all__ = ['WanVACEImageReference', 'WanVACEVideoRepainting', 'WanVACEVideoEdit', 
+           'WanVACEVideoExtension', 'WanVACEVideoOutpainting']
