+++ vace/video_outpainting.py
@@ -1,5 +1,5 @@
 """
-Wan VACE Multi-Image Reference Node for ComfyUI
+Wan VACE Video Outpainting Node for ComfyUI
 """
 
 import os
@@ -16,7 +16,7 @@ import pathlib
 from datetime import datetime
 
 # Import the base class and COMFYUI_AVAILABLE flag
-from .wan_base import WanAPIBase, COMFYUI_AVAILABLE
+from ..core.base import WanAPIBase, COMFYUI_AVAILABLE
 
 # Try to import folder_paths if available
 try:
@@ -24,27 +24,18 @@ try:
 except ImportError:
     pass
 
-class WanVACEImageReference(WanAPIBase):
-    """Node for multi-image reference using Wan VACE model"""
+class WanVACEVideoOutpainting(WanAPIBase):
+    """Node for video outpainting using Wan VACE model"""
     
     # Define available Wan VACE models
     MODEL_OPTIONS = [
         "wan2.1-vace-plus"    # Professional Edition
     ]
     
-    # Define video resolutions
-    RESOLUTION_OPTIONS = [
-        "1280*720",           # 16:9 aspect ratio (default)
-        "720*1280",           # 9:16 aspect ratio
-        "960*960",            # 1:1 aspect ratio
-        "832*1088",           # 3:4 aspect ratio
-        "1088*832"            # 4:3 aspect ratio
-    ]
-    
     def __init__(self):
         super().__init__()
         # Use the centralized API endpoint from the base class
-        # To use Mainland China region, modify API_ENDPOINT_POST_VIDEO in wan_base.py
+        # To use Mainland China region, modify API_ENDPOINT_POST_VIDEO in core/base.py
         self.api_url = self.API_ENDPOINT_POST_VIDEO
     
     @classmethod
@@ -70,22 +61,37 @@ class WanVACEImageReference(WanAPIBase):
                 }),
                 "prompt": ("STRING", {
                     "multiline": True,
-                    "default": "Generate a video based on the provided reference images"
+                    "default": "Outpaint the video with the following description"
                 }),
-                "ref_images_url": ("STRING", {
-                    "multiline": True,
+                "video_url": ("STRING", {
                     "default": "",
-                    "tooltip": "Newline-separated URLs for reference images"
+                    "tooltip": "URL of the input video"
                 })
             },
             "optional": {
-                "obj_or_bg": ("STRING", {
-                    "multiline": True,
-                    "default": "",
-                    "tooltip": "Newline-separated values (obj/bg) corresponding to ref_images_url"
+                "top_scale": ("FLOAT", {
+                    "default": 1.0,
+                    "min": 1.0,
+                    "max": 2.0,
+                    "step": 0.1
                 }),
-                "size": (cls.RESOLUTION_OPTIONS, {
-                    "default": "1280*720"
+                "bottom_scale": ("FLOAT", {
+                    "default": 1.0,
+                    "min": 1.0,
+                    "max": 2.0,
+                    "step": 0.1
+                }),
+                "left_scale": ("FLOAT", {
+                    "default": 1.0,
+                    "min": 1.0,
+                    "max": 2.0,
+                    "step": 0.1
+                }),
+                "right_scale": ("FLOAT", {
+                    "default": 1.0,
+                    "min": 1.0,
+                    "max": 2.0,
+                    "step": 0.1
                 }),
                 "seed": ("INT", {
                     "default": 0,
@@ -106,31 +112,22 @@ class WanVACEImageReference(WanAPIBase):
     FUNCTION = "generate"
     CATEGORY = "Ru4ls/Wan/VACE"
     
-    def generate(self, model, prompt, ref_images_url, obj_or_bg="", size="1280*720", 
-                 seed=0, prompt_extend=False, watermark=False, output_dir="./videos"):
+    def generate(self, model, prompt, video_url, top_scale=1.0, bottom_scale=1.0, 
+                 left_scale=1.0, right_scale=1.0, seed=0, prompt_extend=False, 
+                 watermark=False, output_dir="./videos"):
         
         # Check API key
         self.check_api_key()
-        
-        # Validate required inputs
-        if not ref_images_url or not ref_images_url.strip():
-            raise ValueError("Reference images URLs are required for image reference function")
-        
-        # Handle ref_images_url as a list
-        ref_images_list = [url.strip() for url in ref_images_url.split('\n') if url.strip()]
-        if not ref_images_list:
-            raise ValueError("At least one valid reference image URL is required")
         
         # Prepare API payload
         payload = {
             "model": model,
             "input": {
-                "function": "image_reference",
+                "function": "video_outpainting",
                 "prompt": prompt,
-                "ref_images_url": ref_images_list
+                "video_url": video_url
             },
             "parameters": {
-                "size": size,
                 "prompt_extend": prompt_extend,
                 "watermark": watermark
             }
@@ -139,25 +136,16 @@ class WanVACEImageReference(WanAPIBase):
         # Add seed if provided
         if seed > 0:
             payload["parameters"]["seed"] = seed
-        
-        # Handle obj_or_bg as a list
-        if obj_or_bg:
-            obj_or_bg_list = [item.strip() for item in obj_or_bg.split('\n') if item.strip()]
-            if obj_or_bg_list:
-                # Validate that obj_or_bg_list has the same length as ref_images_list
-                if len(obj_or_bg_list) != len(ref_images_list):
-                    raise ValueError("obj_or_bg list must have the same length as ref_images_url list")
-                payload["parameters"]["obj_or_bg"] = obj_or_bg_list
-        else:
-            # If obj_or_bg is not provided, automatically generate it:
-            # - For single image: default to ["obj"]
-            # - For multiple images: assume last image is background, rest are entities
-            if len(ref_images_list) == 1:
-                payload["parameters"]["obj_or_bg"] = ["obj"]
-            else:
-                # For multiple images, assume last is bg, rest are obj
-                obj_or_bg_auto = ["obj"] * (len(ref_images_list) - 1) + ["bg"]
-                payload["parameters"]["obj_or_bg"] = obj_or_bg_auto
+            
+        # Add scale parameters if not default
+        if top_scale != 1.0:
+            payload["parameters"]["top_scale"] = top_scale
+        if bottom_scale != 1.0:
+            payload["parameters"]["bottom_scale"] = bottom_scale
+        if left_scale != 1.0:
+            payload["parameters"]["left_scale"] = left_scale
+        if right_scale != 1.0:
+            payload["parameters"]["right_scale"] = right_scale
         
         # Set headers according to DashScope documentation
         headers = {
@@ -222,7 +210,7 @@ class WanVACEImageReference(WanAPIBase):
         import time
         
         # URL for querying task results
-        # To use Mainland China region, modify API_ENDPOINT_GET in wan_base.py
+        # To use Mainland China region, modify API_ENDPOINT_GET in core/base.py
         query_url = self.API_ENDPOINT_GET.format(task_id=task_id)
         
         headers = {
@@ -254,7 +242,7 @@ class WanVACEImageReference(WanAPIBase):
                         
                         # Create a unique filename for the video
                         timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-                        video_filename = f"wan_vace_image_reference_{timestamp}.mp4"
+                        video_filename = f"wan_vace_video_outpainting_{timestamp}.mp4"
                         
                         # Handle output directory based on ComfyUI availability
                         if COMFYUI_AVAILABLE and not output_dir.startswith(("./", "/")):
